backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
redis==5.2.0
aiofiles==24.1.0
pyotp==2.9.0
qrcode==8.0
cryptography==43.0.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
email-validator==2.2.0
